Create BreakPractice
For Loops have ALWAYS been an issue for me when learning Python. This is my 3rd or 4th attempt learning with a new course/tutor. Working to understand break statements so that I can continue this list of exercises on https://pynative.com/python-if-else-and-for-loop-exercise-with-solutions/#h-exercise-1-print-first-10-natural-numbers-using-while-loop . Any and all Help is Welcomed.